stateless renderer implementation base
