+++ v3/src/renderer/_webgl/batches/shape/ShapeBatch.js
