+++ v3/src/renderer/_webgl/batches/shape/earcut.js
