+++ v3/src/renderer/_webgl/batches/sprite/FragmentShader.js
