+++ v3/src/renderer/_webgl/batches/sprite/SpriteBatch.js
