+++ v3/src/renderer/_webgl/batches/sprite/VertexShader.js
