+++ v3/src/renderer/_webgl/utils/CreateEmptyTexture.js
